// frontend/src/types/api.d.ts
export interface paths {
    "/events": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["event/get/events"];
        put?: never;
        post: operations["event/post/events"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/events/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["event/get/events_by_id"];
        put: operations["event/put/events_by_id"];
        post?: never;
        delete: operations["event/delete/events_by_id"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/people": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["person/get/people"];
        put?: never;
        post: operations["person/post/people"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/people/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["person/get/people_by_id"];
        put: operations["person/put/people_by_id"];
        post?: never;
        delete: operations["person/delete/people_by_id"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/price": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["price/get/price"];
        put: operations["price/put/price"];
        post?: never;
        delete: operations["price/delete/price"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/tarif-distances": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["tarif-distance/get/tarif_distances"];
        put?: never;
        post: operations["tarif-distance/post/tarif_distances"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/tarif-distances/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["tarif-distance/get/tarif_distances_by_id"];
        put: operations["tarif-distance/put/tarif_distances_by_id"];
        post?: never;
        delete: operations["tarif-distance/delete/tarif_distances_by_id"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: never;
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    "event/get/events": {
        parameters: {
            query?: {
                fields?: ("Name" | "Distance" | "Location" | "createdAt" | "updatedAt" | "publishedAt")[];
                filters?: {
                    [key: string]: unknown;
                };
                _q?: string;
                pagination?: {
                    withCount?: boolean;
                } & ({
                    page: number;
                    pageSize: number;
                } | {
                    start: number;
                    limit: number;
                });
                sort?: ("Name" | "Distance" | "Location" | "createdAt" | "updatedAt" | "publishedAt") | ("Name" | "Distance" | "Location" | "createdAt" | "updatedAt" | "publishedAt")[] | {
                    [key: string]: "asc" | "desc";
                } | {
                    [key: string]: "asc" | "desc";
                }[];
                populate?: "*" | ("contacts" | "Budget") | ("contacts" | "Budget")[];
                status?: "draft" | "published";
                hasPublishedVersion?: boolean | ("true" | "false");
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
// ...
    };
// ...
}